+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Generation of scope directory
 * Generation of x.provider.json file
 * Workaround for document type flag
+* `isFileQuiet` function
 
 ### Changed
 * Updated parser dependency to v0.0.12
@@ -13,9 +14,12 @@
 * Updated compile and lint tests to use `path.join` instead of string paths
 * Play command outputs build files to `<playground>/build` directory
 * Playground no longer needs to have the same name as the profile within it
+* Playground now works with playscripts in `superface/play` folder
+* Playground can now work with scoped profiles and maps
 
 ### Removed
 * Document type flag: `--documentType` (`-t`)
+* Unnecessary fixture playground
 
 ## [0.0.2]